feat!: reduce project scope to specification compliance verification
Open Source project maintainers are accountable for their software
products even though they are not paid for the work. Further, the
risk in healthcare applications is particularly high so that users
of openEHR4J need to justify its usage. By reducing the project
scope, it has an inherently safe design which eliminates the main
risks for the maintainer and clients.